+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Suggest last used value in measurable
 - Upgrade dependencies
+- More responsiveness in measurement creation
 
 ## [0.8.202] - 2022-12-01
 ### Added: